Add files via upload
Lots of the classes are getting very interlinked / interdependent. If I want to ship this to an app it can't be all at once, so I've backed off on a few of the extra features and focused on turning this into a stand-alone class. There's some homeless code in here that will be useful later, and some well-commented zombie code, but overall things are much neater and self-contained.