Add files via upload
TownNameGenerator class has been deprecated. It's functionality has been rolled into LocationNameGenerator, which also has methods for generating names for geographical and man-made landmarks.
-Added support for individual names/linguistic influences. These are generated by the WordGen class.

WordGen is a new class that will eventually replace all functionality of the  "CharacterNameGenerator" class. It currently exists as a single class but may be separated into several smaller ones depending on where the project goes in the future.
-Words and names are generated procedurally, rather than pulled from sets. Each language uses different rules, as well as sets of vowels, consonants, and chunks to generate words/names that with a character unique to that language.
-Currently supports the Elf, Orc, and Common language. (Common needs some tweaking).